# docs/swagger.yaml
definitions:
  controllers.Response:
    properties:
      message: {}
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.Notif:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      relation_id:
        type: integer
      type:
        type: string
      user_id:
        type: integer
    type: object
  models.NotifRequest:
    properties:
      message:
        type: string
      relation_id:
        type: integer
      type:
        type: string
      user_id:
        type: integer
    required:
    - message
    - relation_id
    - type
    - user_id
    type: object
...
  models.UserRegister:
    properties:
      email:
        type: string
      first_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role:
        enum:
        - client
        - contractor
        type: string
    required:
    - email
    - first_name
    - phone_number
    - role
    type: object
  models.VerifyRequest:
    properties:
      code:
        type: string
      phone_number:
        type: string
    type: object
info:
  contact:
    email: muhtorhongofurov@gmail.com
    name: Muxtorxon Gofurov
    url: https://github.com/muhtorkhon
  description: Tender Management Golang REST API
  title: Tender Management REST API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Allows a user to log in using email and password. If valid, returns
        a JWT token.
      parameters:
      - description: Login Credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored the category
          schema:
            $ref: '#/definitions/controllers.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.Response'
        "401":
          description: 'Unauthorized: Invalid credentials'
          schema:
            $ref: '#/definitions/controllers.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.Response'
      summary: Login a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user by providing phone number, email, and password.
      parameters:
      - description: User Registration Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created the user
          schema:
            $ref: '#/definitions/controllers.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/controllers.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.Response'
      summary: Create a new user
      tags:
      - auth
  /auth/verify:
    post:
      consumes:
      - application/json
      description: Verifies the user's phone number using the code sent earlier. If
        valid, activates the user.
      parameters:
      - description: Phone number and code verification data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User verified and activated successfully
          schema:
            type: string
        "400":
          description: Invalid or expired code
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Verify phone number with the code
      tags:
      - auth
  /client/notifs/{user_id}/{relation_id}:
    get:
      description: Client uchun o‘ziga tegishli xabarni olish
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Relation ID
        in: path
        name: relation_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Notif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.Response'
        "404":
          description: Notif not found
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Get Client Notification
      tags:
      - Notifications
  /contractor/notifs/{user_id}/{relation_id}:
    get:
      description: Contractor uchun o‘ziga tegishli xabarni olish
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Relation ID
        in: path
        name: relation_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Notif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.Response'
        "404":
          description: Notif not found
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Get Contractor Notification
      tags:
      - Notifications
  /notifs:
    post:
      consumes:
      - application/json
      description: Yangi xabar yaratish (Client yoki Contractor uchun)
      parameters:
      - description: Notification Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.NotifRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Notif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.Response'
      summary: Create Notification
      tags:
      - Notifications
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"